requirements.txt
alembic==1.16.5
blinker==1.9.0
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fonttools==4.59.2
greenlet==3.2.4
gunicorn==23.0.0
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.0
networkx==3.5
numpy==2.2.1
nvidia-nccl-cu12==2.27.7
opencv-python==4.10.0.84
packaging==25.0
pandas==2.2.3
pillow==11.0.0
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-image==0.25.0
scipy==1.14.1
six==1.17.0
SQLAlchemy==2.0.36
tifffile==2025.8.28
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
xgboost==2.1.3
